requirements.txt
# GameChat AI Backend Dependencies
# This file is used for both development and production environments

# Core dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.82.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
upstash-vector==0.8.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Google Cloud Storage
google-cloud-storage>=2.13.0

# Authentication and Security (required for production)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate limiting and caching
redis>=5.0.0
slowapi>=0.1.9

# Development utilities
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality and type checking
mypy>=1.16.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Documentation generation
sphinx>=7.1.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0

# Production monitoring (optional - uncomment if needed)
# sentry-sdk[fastapi]>=1.38.0
